+++ composeApp/src/main/kotlin/com/example/a62550_foodapp/di/AppModule.kt
@@ -95,7 +95,7 @@ val appModule = module {
 
     viewModel{ ShoppingListViewModel(get()) }
 
-    viewModel { (shoppingListId: Int) ->
+    viewModel { (shoppingListId: Long) ->
         ShoppingListDetailsViewModel(
             shoppingListId = shoppingListId,
             shoppingListItemGroupDao = get(),
